requirements.txt
github-readme-terminal
python-dotenv